Component based strutcure with updated imports
