+++ app/main.js
@@ -1,5 +1,6 @@
 
 import { head, title } from "../@butility/dom/html.js";
+import { loadWebFont } from "../@butility/style/style.js";
 
 const queryString = window.location.search;
 const params = new URLSearchParams(queryString);
@@ -9,6 +10,11 @@ if (params.get("page") === "ip") {
 } else {
     await import("./form.js");
 }
+
+loadWebFont({
+    family: "Roboto",
+    elements: [ "body" ],
+});
 
 head(
     title("Butility example")
